🔨+🎨 refactor interfaces
